Segregated login and auth controllers
